Added additional files beyond src folder
Added files required for running in WebStorm
